initial commit for children vnf handlers
